refactor(Utils): separate some translate source
