+++ src/components/MachineTPanel.js
@@ -1,9 +1,9 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as Utils from "../Utils.js";
 import * as React from "react";
 import * as Widget from "./Widget.js";
 import * as Caml_obj from "rescript/lib/es6/caml_obj.js";
+import * as TranSource from "../TranSource.js";
 import * as Belt_Option from "rescript/lib/es6/belt_Option.js";
 import * as Caml_option from "rescript/lib/es6/caml_option.js";
 
@@ -72,7 +72,7 @@ function MachineTPanel(props) {
                                       audio.src = "";
                                       return ;
                                     }
-                                    audio.src = Utils.Baidu.textToSpeech(text);
+                                    audio.src = TranSource.Baidu.textToSpeech(text);
                                     audio.play();
                                   }));
                           }),
@@ -93,4 +93,4 @@ var make = MachineTPanel;
 export {
   make ,
 }
-/* Utils Not a pure module */
+/* react Not a pure module */
