+++ src/components/Widget.js
@@ -1,7 +1,7 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as Utils from "../Utils.js";
 import * as React from "react";
+import * as Functions from "../Functions.js";
 import * as Caml_option from "rescript/lib/es6/caml_option.js";
 
 function shouldDelay(spanning, delay) {
@@ -29,7 +29,7 @@ function Widget$Loading(props) {
   var setSpinning = match[1];
   React.useEffect((function () {
           if (loading) {
-            var match = Utils.Lib.debounce(delay, (function () {
+            var match = Functions.debounce(delay, (function () {
                     setSpinning(function (param) {
                           return true;
                         });
@@ -334,4 +334,4 @@ export {
   EyeSlash ,
   Search ,
 }
-/* Utils Not a pure module */
+/* react Not a pure module */
