+++ src/components/Widget.res
@@ -1,4 +1,4 @@
-open Utils.Lib
+open Functions
 
 module Loading = {
   let shouldDelay = (spanning, delay) => {
