+++ src/content_scripts/ContentApp.js
@@ -1,9 +1,9 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as Utils from "../Utils.js";
 import * as React from "react";
 import * as Widget from "../components/Widget.js";
 import * as FavButton from "../components/FavButton.js";
+import * as Functions from "../Functions.js";
 import * as TranslateHook from "../hooks/TranslateHook.js";
 import * as TranslateResult from "../components/TranslateResult.js";
 
@@ -44,7 +44,7 @@ function ContentApp(props) {
               return ;
             }
             var range = selection.getRangeAt(0);
-            var sl = Utils.getSourceLang(text);
+            var sl = Functions.getSourceLang(text);
             setSrcLang(function (param) {
                   return sl;
                 });
@@ -60,7 +60,7 @@ function ContentApp(props) {
             setLeft(function (_p) {
                   return left.toString() + "px";
                 });
-            var match = Utils.Lib.debounce(200, (function () {
+            var match = Functions.debounce(200, (function () {
                     setOpactity(function (_p) {
                           return "1";
                         });
@@ -154,4 +154,4 @@ export {
   panelDelay ,
   make ,
 }
-/* Utils Not a pure module */
+/* react Not a pure module */
