+++ src/content_scripts/ContentApp.res
@@ -1,6 +1,6 @@
 open Widget
 open Utils
-open Utils.Lib
+open Functions
 open TranslateResult
 
 open Common.Webapi
