+++ src/hooks/RecordHook.js
@@ -1,8 +1,8 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as Utils from "../Utils.js";
 import * as React from "react";
 import * as Caml_obj from "rescript/lib/es6/caml_obj.js";
+import * as Functions from "../Functions.js";
 
 function useRecord(recordType) {
   var match = React.useState(function () {
@@ -68,7 +68,7 @@ function useRecord(recordType) {
   var onSearch = function (val) {
     if (val !== "") {
       var rs = records.filter(function (item) {
-            return Utils.includeWith(item.text, val);
+            return Functions.includeWith(item.text, val);
           });
       return setRecords(function (param) {
                   return rs;
@@ -137,4 +137,4 @@ function useRecord(recordType) {
 export {
   useRecord ,
 }
-/* Utils Not a pure module */
+/* react Not a pure module */
