+++ src/options/Login.js
@@ -1,8 +1,8 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as Utils from "../Utils.js";
 import * as React from "react";
 import * as Widget from "../components/Widget.js";
+import * as Functions from "../Functions.js";
 import * as RescriptReactRouter from "@rescript/react/src/RescriptReactRouter.js";
 
 function Login(props) {
@@ -25,7 +25,7 @@ function Login(props) {
   var setPasswordVisible = match$2[1];
   var passwordVisible = match$2[0];
   var handleSubmit = async function () {
-    var val = await Utils.Lib.fetchByHttp("/login", "post", {
+    var val = await Functions.fetchByHttp("/login", "post", {
           username: username,
           password: password
         });
@@ -107,4 +107,4 @@ var make = Login;
 export {
   make ,
 }
-/* Utils Not a pure module */
+/* react Not a pure module */
