+++ src/options/TranslateService.js
@@ -1,8 +1,8 @@
 // Generated by ReScript, PLEASE EDIT WITH CARE
 
-import * as Utils from "../Utils.js";
 import * as React from "react";
 import * as Widget from "../components/Widget.js";
+import * as TranSource from "../TranSource.js";
 
 function TranslateService(props) {
   var match = React.useState(function () {
@@ -53,7 +53,7 @@ function TranslateService(props) {
     chrome.storage.local.set({
           baiduKey: config
         });
-    var br = await Utils.Baidu.translate("hello world");
+    var br = await TranSource.Baidu.translate("hello world");
     if (br.TAG === "Ok") {
       await chrome.storage.local.set({
             baiduKey: config
@@ -142,4 +142,4 @@ var make = TranslateService;
 export {
   make ,
 }
-/* Utils Not a pure module */
+/* react Not a pure module */
